+++ examples/readme/src/main.rs
@@ -2,9 +2,9 @@ use std::{error::Error, net::SocketAddr, time::Duration};
 
 use clap::Parser;
 use itertools::Itertools;
-use sunspec::client::{
-    tokio_modbus::{discover_models, read_model},
-    Config,
+use sunspec::{
+    client::{AsyncClient, Config},
+    models::{model1::Model1, model103::Model103},
 };
 use tokio::time::sleep;
 use tokio_modbus::{client::tcp::connect_slave, Slave};
@@ -18,11 +18,14 @@ struct Args {
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn Error>> {
     let args = Args::parse();
-    let mut ctx = connect_slave(args.addr, Slave(args.device_id)).await?;
 
-    let cfg = Config::default();
-    let models = discover_models(&mut ctx, &cfg).await?.models;
-    let m1 = read_model(&mut ctx, models.m1, &cfg).await?;
+    let mut client = AsyncClient::new(
+        connect_slave(args.addr, Slave(args.device_id)).await?,
+        Config::default(),
+    );
+
+    client.discover_models().await?;
+    let m1: Model1 = client.read_model().await?;
 
     println!("Manufacturer: {}", m1.mn);
     println!("Model: {}", m1.md);
@@ -31,7 +34,8 @@ async fn main() -> Result<(), Box<dyn Error>> {
 
     println!(
         "Supported models: {}",
-        models
+        client
+            .models
             .supported_model_ids()
             .iter()
             .map(|id| id.to_string())
@@ -39,7 +43,7 @@ async fn main() -> Result<(), Box<dyn Error>> {
     );
 
     loop {
-        let m103 = read_model(&mut ctx, models.m103, &cfg).await?;
+        let m103: Model103 = client.read_model().await?;
         let w = m103.w as f32 * 10f32.powf(m103.w_sf.into());
         let wh = m103.wh as f32 * 10f32.powf(m103.wh_sf.into());
         println!("{:12.3} kWh {:9.3} kW", wh / 1000.0, w / 1000.0,);