+++ CHANGELOG.md
@@ -7,12 +7,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add `AsyncClient` which provides a more ergonomic API
+
 ### Changed
 
-- `ModelAddr` and `PointRef` now implement `Copy` and `Clone` and are passed by
-  value and not by reference.
+- `ModelAddr` and `PointRef` now implement `Copy` and `Clone` and are passed
+  by value and not by reference.
 - Move all client specific code into `client` module.
 - Rename `PointDef` to `Point`.
+
+### Removed
+
+- Remove old client functions
 
 ## [0.6.1] - 2024-10-18